-windows 64-bit build fixes
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@9649 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/TransportSelector.cxx
@@ -268,7 +268,7 @@ TransportSelector::addTransportInternal(std::auto_ptr<Transport> autoTransport)
    }
 
    mTransports.push_back(transport);
-   transport->setKey(mTransports.size());
+   transport->setKey((unsigned int)mTransports.size());
 }
 
 void
